+++ electron/electron-env.d.ts
@@ -17,6 +17,11 @@ declare namespace NodeJS {
      */
     DIST: string
     /** /dist/ or /public/ */
-    PUBLIC: string
+    VITE_PUBLIC: string
   }
+}
+
+// Used in Renderer process, expose in `preload.ts`
+interface Window {
+  ipcRenderer: import('electron').IpcRenderer
 }